+++ projects/strct/src/public-api.ts
@@ -47,6 +47,7 @@ export * from './lib/pagination/pagination';
 export * from './lib/section-menu/section-menu';
 export * from './lib/command-palette/command-palette';
 export * from './lib/kbd/kbd';
+export * from './lib/searchbox/searchbox';
 
 // Forms
 export * from './lib/forms/field';